RethinkDB and Connection object exposed right
THIS IS A BREAKING CHANGE!

Now the connection object is exposed as `this.rethinkdbConn` and the
RethinkDB library is exposed as `this.rethinkdb`. This way to expose it
is more in line with how it's used in the examples given all around the
documentation of RethinkDB.

It changed the names of how is exposed on
`server.plugins['hapi-rethinkdb']` as well. Now instead of using
`client` it's being used `connection`; `library` was kept the same way.